move default_category_colors to new lib/colors.ts
