+++ src/lib/colors.ts
@@ -1,0 +1,15 @@
+// color values have to be in hex format since that's the only format
+// <input type="color"> supports
+// https://developer.mozilla.org/en-US/docs/Web/HTML/Element/input/color#value
+export const default_category_colors: Record<string, string> = {
+  'diatomic-nonmetal': `#ff8c00`, // darkorange
+  'noble-gas': `#9932cc`, // darkorchid
+  'alkali-metal': `#006400`, // darkgreen
+  'alkaline-earth-metal': `#483d8b`, // darkslateblue
+  metalloid: `#b8860b`, // darkgoldenrod
+  'polyatomic-nonmetal': `#a52a2a`, // brown
+  'transition-metal': `#571e6c`,
+  'post-transition-metal': `#938d4a`,
+  lanthanide: `#58748e`,
+  actinide: `#6495ed`, // cornflowerblue
+}
